+++ routes/web.php
@@ -2,6 +2,7 @@
 
 declare(strict_types=1);
 
+use App\Enums\PermissionType;
 use App\Http\Controllers\CommentController;
 use App\Http\Controllers\FeatureController;
 use App\Http\Controllers\ProfileController;
@@ -21,13 +22,15 @@ Route::middleware('auth')->group(function () {
             return Inertia::render('Dashboard');
         })->name('dashboard');
 
-        Route::resource('feature', FeatureController::class);
+        Route::resource('feature', FeatureController::class)->except('index', 'show')->middleware('can:'.PermissionType::ManageFeature->value);
+        Route::get('/feature', [FeatureController::class, 'index'])->name('feature.index');
+        Route::get('/feature/{feature}', [FeatureController::class, 'show'])->name('feature.show');
 
         Route::post('/feature/{feature}/upvote', [UpvoteController::class, 'store'])->name('upvote.store');
         Route::delete('/upvote/{feature}', [UpvoteController::class, 'destroy'])->name('upvote.destroy');
 
-        Route::post('/feature/{feature}/comments', [CommentController::class, 'store'])->name('comment.store');
-        Route::delete('/comment/{comment}', [CommentController::class, 'destroy'])->name('comment.destroy');
+        Route::post('/feature/{feature}/comments', [CommentController::class, 'store'])->name('comment.store')->middleware('can:'.PermissionType::ManageComments->value);
+        Route::delete('/comment/{comment}', [CommentController::class, 'destroy'])->name('comment.destroy')->middleware('can:'.PermissionType::ManageComments->value);
     });
 });
